camp_status: 7 scoped-pragma wins (DrawSp 591->363, FUN_0012B300 897->755, FUN_0012BCE0 449->391, root 703->664, PanelTask 634->614, DrawSteady 1115->1088, DrawHp 767->765)

+++ src/h_snd.c
@@ -403,7 +403,7 @@ void func_00108bc0(void)
         s32 status;
         s16* state;
 
-        for (i = 2; (s16)i < HSND_SLOT_COUNT; )
+        for (i = 2; i < HSND_SLOT_COUNT; )
         {
             channel = &sChannels[i];
             if (channel->active != false)
